Fork the git plugin into deprecated-git
Splitting out a current copy of the code, in advance of a radical
simplification as described in [1].

This patch was generated by copying the `plugins/git` directory, and
then modifying instances of the string `"/git/"` to
`"/deprecated-git/"`. Most notably, the file
`plugins/deprecated-git/loadRepositoryTest.sh` has been updated so that
it's example data file points to a data file in `deprecated-git`, and
the corresponding `loadAndPrintGitRepository` `bin/` target has been
forked so there is now a version specific to the deprecated plugin.

Test plan: `yarn travis --full` works, as does
`src/plugins/deprecated-git/loadRepositoryTest.sh`.

[1]: https://github.com/sourcecred/sourcecred/issues/627#issuecomment-413435447