readme: reflow to one-line-per-sentence
Summary:
This patch does not change the observable content of the README.
Instead, it reflows the text to insert a newline after each sentence.
This approach makes it much easier to read diffs and follow blame.

I speak from experience: having maintained medium-sized prose projects
with multiple collaborators, the diffs and blame get ugly/useless if you
insist on wrapping at a fixed column width. Sentences are a natural
place to break instead, and very convenient: operations like permuting
and deleting sentences, which are common in prose editing, are now
trivial in Vim/Emacs/etc. In systems like Markdown and TeX, where
breaking lines thus “just works”, it’s a no-brainer.

The fact that line length is unbounded is—again, in my experience—not
really a problem. Having a line length cap is nice for code, because
code looks bad when it wraps (not least due to large leading
indentation). But prose looks fine when it wraps, so just `:set wrap` in
Vim/etc and you’re off.

See also <http://rhodesmill.org/brandon/2012/one-sentence-per-line/>,
but I’ve been convinced (in part by people who know better than me, and
in part from my own experience) that actual semantic linefeeds as
described in that post are inferior to breaking at each sentence,
primarily because it is awkward to read, and also because it makes
deciding where to insert the breaks a non-trivial task.

Also, <https://www.xkcd.com/1285/>.