Factor out WeightSlider and DirectionalitySlider
This commit factors the weight sliders used for both node and edge
weights into a shared WeightSlider component, and factors out the
direction slider used for edge weights into a DirectionalitySlider.

Both of these components are tested. This is a step towards #604.

Test plan:
The specific behaviors of the sliders are well tested. Since the weight
config as a whole is not tested, I manually verified by messing with the
weights that node weights, edge weights, and edge directionality all
affects the cred distribution as anticipated.

Note to reviewers: Heretofore, when writing jest tests, I have scoped
all of the tests in a file in a describe block with the path to that
file. I'm finding this practice annoying and brittle, since when we move
files we rarely remember to update the description, so as of this commit
I've stopped doing that. If we decide we're better off without those
describes, then I'll remove them elsewhere in the codebase later.